+++ scripts/build-deps.ps1
@@ -111,7 +111,7 @@ function buildSentryCocoa()
 
     try
     {
-        make build-xcframework-sentryobjc-dynamic SDKS=iphoneos,macosx
+        make build-xcframework-sentryobjc-dynamic SDKS=iphoneos,iphonesimulator,macosx
 
         if ($LASTEXITCODE -ne 0)
         {
@@ -139,15 +139,16 @@ function buildSentryCocoa()
 
     New-Item $iosOutDir -ItemType Directory > $null
 
-    Copy-Item "$xcframeworkPath/ios-arm64/SentryObjC.framework" -Destination "$iosOutDir/SentryObjC.framework" -Recurse
-
     New-Item "$iosOutDir/SentryObjC.embeddedframework" -ItemType Directory > $null
     Copy-Item "$xcframeworkPath/ios-arm64/SentryObjC.framework" -Destination "$iosOutDir/SentryObjC.embeddedframework/SentryObjC.framework" -Recurse
+
+    Copy-Item $xcframeworkPath -Destination "$iosOutDir/SentryObjC.xcframework" -Recurse
 
     Push-Location $iosOutDir
     try
     {
         zip -r "SentryObjC.embeddedframework.zip" "SentryObjC.embeddedframework"
+        zip -r "SentryObjC.xcframework.zip" "SentryObjC.xcframework"
     }
     finally
     {
@@ -155,6 +156,7 @@ function buildSentryCocoa()
     }
 
     Remove-Item "$iosOutDir/SentryObjC.embeddedframework" -Recurse -Force
+    Remove-Item "$iosOutDir/SentryObjC.xcframework" -Recurse -Force
 
     Write-Host "Successfully built Sentry Cocoa for iOS"