+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Features
+
+- Add support for iOS Simulator builds on UE 5.5 and newer ([#1560](https://github.com/getsentry/sentry-unreal/pull/1560))
+
 ## 1.22.0
 
 ### Features