+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 - Bump Java SDK from v8.54.0 to v8.55.0 ([#1560](https://github.com/getsentry/sentry-unreal/pull/1560))
   - [changelog](https://github.com/getsentry/sentry-java/blob/main/CHANGELOG.md#8550)
   - [diff](https://github.com/getsentry/sentry-java/compare/8.54.0...8.55.0)
+- Bump Android Gradle Plugin from v6.20.0 to v6.21.0 ([#1562](https://github.com/getsentry/sentry-unreal/pull/1562))
+  - [changelog](https://github.com/getsentry/sentry-android-gradle-plugin/blob/main/CHANGELOG.md#6210)
+  - [diff](https://github.com/getsentry/sentry-android-gradle-plugin/compare/6.20.0...6.21.0)
 
 ## 1.22.0